Now can render UTF-8 text Well with tons of crap but it is a thing

+++ src/include/types.h
@@ -5571,7 +5571,7 @@ struct fontchar {
 #else
 	s32 kerningindex;
 #endif
-	u8 *pixeldata;
+	uint16_t *pixeldata;
 };
 
 struct font {
